[8. String to Integer (atoi)][Compile Error]committed by Anthony

+++ 8-String-to-Integer-(atoi)/solution.java
@@ -42,7 +42,7 @@ public class Solution {
         // rule out the string > 2^31 - 1 which is 2147483647 or string < -2^31 which is -2147483648
         
         if (initial == 45) {
-            result = Long.valueOf(str.substring(1, str.length())) * (-1);
+            result = (int)(Long.valueOf(str.substring(1, str.length())) * (-1));
             if (result < -2147483648) {
                 return 0;
             } else {
